requirements.txt
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Session==0.5.0
SQLAlchemy==2.0.23
cryptography==42.0.5
pyotp==2.9.0
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
python-dotenv==1.0.0
qrcode==7.4.2